+++ passwd/passwd_test.go
@@ -61,8 +61,13 @@ func (s *PasswdSuite) TestGenPasswordVariations(c *C) {
 	})
 }
 
-func (s *PasswdSuite) TestEncrypt(c *C) {
+func (s *PasswdSuite) TestHash(c *C) {
 	hp, err := Encrypt("Test123", "ABCD1234ABCD1234")
+
+	c.Assert(hp, NotNil)
+	c.Assert(err, IsNil)
+
+	hp, err = Hash("Test123", "ABCD1234ABCD1234")
 
 	c.Assert(hp, NotNil)
 	c.Assert(err, IsNil)
@@ -78,20 +83,20 @@ func (s *PasswdSuite) TestEncrypt(c *C) {
 	c.Assert(Check("TEST", "ABCD1234ABCD1234", "0000000000000000000000"), Equals, false)
 }
 
-func (s *PasswdSuite) TestEncryptErrors(c *C) {
+func (s *PasswdSuite) TestHashErrors(c *C) {
 	var err error
 
-	_, err = Encrypt("", "ABCD1234ABCD1234")
+	_, err = Hash("", "ABCD1234ABCD1234")
 
 	c.Assert(err, NotNil)
 	c.Assert(err.Error(), Equals, "Password can't be empty")
 
-	_, err = Encrypt("Test123", "")
+	_, err = Hash("Test123", "")
 
 	c.Assert(err, NotNil)
 	c.Assert(err.Error(), Equals, "Pepper can't be empty")
 
-	_, err = Encrypt("Test123", "ABCD1234ABCD12")
+	_, err = Hash("Test123", "ABCD1234ABCD12")
 
 	c.Assert(err, NotNil)
 	c.Assert(err.Error(), Equals, "Pepper have invalid size")
@@ -101,9 +106,9 @@ func (s *PasswdSuite) TestEncryptErrors(c *C) {
 	c.Assert(ok, Equals, false)
 }
 
-func (s *PasswdSuite) BenchmarkEncrypt(c *C) {
+func (s *PasswdSuite) BenchmarkHash(c *C) {
 	for i := 0; i < c.N; i++ {
-		Encrypt("Test123", "ABCD1234ABCD1234")
+		Hash("Test123", "ABCD1234ABCD1234")
 	}
 }