+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `[secstr]` New package for working with protected (secure) strings
 * `[req]` Method `Query.String()` renamed to `Query.Encode()`
 * `[passwd]` Added method `GenPasswordVariations` for generating password variations with possible typos fixes
+* `[passwd]` Added methods `HashBytes` and `CheckBytes`
+* `[passwd]` Method `Encrypt` marked as depricated (_use `Hash` method instead_)
+* `[passwd]` Added more usage examples
 
 ### 12.34.0